@@ -11,63 +11,81 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
 from __future__ import annotations
 
-from typing import Union
+from typing import Any, Dict, Optional
 
-from pydantic import AliasChoices, AliasPath, Field
+from pydantic import (
+    AliasChoices,
+    AliasPath,
+    Field,
+)
 from pydantic_settings import (
     BaseSettings,
 )
 
 
-class AzureOpenAICredentials(BaseSettings):
+class DRCredentials(BaseSettings): ...
+
+
+class AzureOpenAICredentials(DRCredentials):
     """LLM credentials auto-constructed using environment variables."""
 
-    api_version: str = Field(
+    api_key: str = Field(
         validation_alias=AliasChoices(
-            AliasPath("MLOPS_RUNTIME_PARAM_OPENAI_API_VERSION", "payload"),
-            "OPENAI_API_VERSION",
+            "OPENAI_API_KEY",
+            AliasPath("MLOPS_RUNTIME_PARAM_OPENAI_API_KEY", "payload", "apiToken"),
         ),
     )
     azure_endpoint: str = Field(
         validation_alias=AliasChoices(
-            AliasPath("MLOPS_RUNTIME_PARAM_OPENAI_API_BASE", "payload"),
             "OPENAI_API_BASE",
+            AliasPath("MLOPS_RUNTIME_PARAM_OPENAI_API_BASE", "payload"),
         )
     )
-    api_key: str = Field(
+    api_version: str | None = Field(
+        default=None,
         validation_alias=AliasChoices(
-            AliasPath("MLOPS_RUNTIME_PARAM_OPENAI_API_KEY", "payload", "apiToken"),
-            "OPENAI_API_KEY",
+            "OPENAI_API_VERSION",
+            AliasPath("MLOPS_RUNTIME_PARAM_OPENAI_API_VERSION", "payload"),
         ),
     )
-    azure_deployment: str = Field(
+    azure_deployment: str | None = Field(
+        default=None,
         validation_alias=AliasChoices(
-            AliasPath("MLOPS_RUNTIME_PARAM_OPENAI_API_DEPLOYMENT_ID", "payload"),
             "OPENAI_API_DEPLOYMENT_ID",
-        )
+            AliasPath("MLOPS_RUNTIME_PARAM_OPENAI_API_DEPLOYMENT_ID", "payload"),
+        ),
     )
 
-    def test(self) -> None:
-        import openai
 
-        try:
-            client = openai.AzureOpenAI(
-                api_key=self.api_key,
-                azure_endpoint=self.azure_endpoint,
-                api_version=self.api_version,
-            )
-            client.chat.completions.create(
-                messages=[{"role": "user", "content": "hello"}],
-                model=self.azure_deployment,
-            )
-        except Exception as e:
-            raise ValueError(
-                f"Unable to run a successful test completion against model '{self.azure_deployment}' "
-                "with provided Azure OpenAI credentials. Please validate your credentials."
-            ) from e
+class GoogleCredentials(DRCredentials):
+    service_account_key: Dict[str, Any] = Field(
+        validation_alias="GOOGLE_SERVICE_ACCOUNT"
+    )
+    region: Optional[str] = Field(default="us-west1", validation_alias="GOOGLE_REGION")
 
 
-LLMCredentials = Union[AzureOpenAICredentials]
+class AWSBedrockCredentials(DRCredentials):
+    aws_access_key_id: str = Field(
+        validation_alias=AliasChoices(
+            "AWS_ACCESS_KEY_ID",
+            AliasPath("MLOPS_RUNTIME_PARAM_AWS_ACCOUNT", "payload", "awsAccessKeyId"),
+        )
+    )
+    aws_secret_access_key: str = Field(
+        validation_alias=AliasChoices(
+            "AWS_SECRET_ACCESS_KEY",
+            AliasPath(
+                "MLOPS_RUNTIME_PARAM_AWS_ACCOUNT", "payload", "awsSecretAccessKey"
+            ),
+        )
+    )
+    aws_session_token: Optional[str] = Field(
+        default=None,
+        validation_alias=AliasChoices(
+            "AWS_SESSION_TOKEN",
+            AliasPath("MLOPS_RUNTIME_PARAM_AWS_ACCOUNT", "payload", "awsSessionToken"),
+        ),
+    )
+    region_name: Optional[str] = Field(default=None, validation_alias="AWS_REGION")
